bot/requirements.txt
Django==4.0.5
python-telegram-bot==13.13